Speculative fix for behind live window triggered 404s
If we were too slow to load a segment, maybe retried a couple of
times, the segment may be deleted from the server. We're behind
the live window. In this case the player will simply error out,
which isn't very user friendly (or how iOS does it).

This is a speculative fix:
- Make sure playlist is up to date in HlsChunkSource, not having
  reached the end of our current copy doesn't mean we shouldn't
  refresh it.
- If we get a 404 for a TsChunk in a live stream, try to reset
  state to trigger selection of a more recent chunk.

This needs testing, has not been confirmed to fix the problem yet.

Signed-off-by: Jonas Larsson <jonas@hallerud.se>