AudioTrack: Fix AV sync drift, change time anchor
AudioTrack has a fundamental issue; it only reads the incoming
media timestamp at the start of playback and then reports current position based
on the number of samples written to AudioTrack plus recorded media start
timestamp. Variations in incoming media timestamps are not considered until they
become really big, 200 ms. 200ms is pretty bad AV sync. Just reducing the hard
coded 200 ms gets messy as the existing logic doesn't take input sample length
into account.
Existing code also makes the assumption that a media timestamp corresponds to
the end of a buffer, and it calculates a bufferStartTime as timestamp -
buffer_duration. At the start of streams this often becomes negative, a fact
that's swept under the rug with
startMediaTimeUs = Math.max(0, bufferStartTime);
So audioTrackStartMediaTimeUs is often wrong and doesn't correct itself.

A problematic example is HLS. After discont (eg stream switch) a new TsExtractor
is created with base time taken from the manifest. There can be a slight delta
between the last timestamp in the previous HLS chunk and the new one (with new
base time). These differences add up and AV sync start to drift.

AudioTrack is just a dump sample sink, not a clock source. Media time is the
actual source, so make it so. Record the timestamp and frame number for every
incoming media buffer and calculate the corresponding media time in
getCurrentPositionUs.

Signed-off-by: Jonas Larsson <jonas@hallerud.se>